requirements.txt
Flask
boto3
pycryptodome
langchain
